extract verification to own struct, put protocol around it and made mock from it which return only true
