+++ src/xcode/ENA/ENA/Source/Client/HTTP Client/HTTPClient.swift
@@ -12,12 +12,10 @@ final class HTTPClient: Client {
 
 	init(
 		serverEnvironmentProvider: ServerEnvironmentProviding,
-		packageVerifier: @escaping SAPDownloadedPackage.Verification = SAPDownloadedPackage.Verifier().verify,
 		session: URLSession = .coronaWarnSession()
 	) {
 		self.serverEnvironmentProvider = serverEnvironmentProvider
 		self.session = session
-		self.packageVerifier = packageVerifier
 	}
 
 	// MARK: - Overrides
@@ -406,7 +404,6 @@ final class HTTPClient: Client {
 
 	private let serverEnvironmentProvider: ServerEnvironmentProviding
 	private let session: URLSession
-	private let packageVerifier: SAPDownloadedPackage.Verification
 	private var fetchDayRetries: [URL: Int] = [:]
 	private var traceWarningPackageDownloadRetries: [URL: Int] = [:]
