+++ src/xcode/ENA/ENA/Source/Client/HTTP Client/__tests__/Helper/HTTPClient+MockNetworkStack.swift
@@ -8,14 +8,12 @@ import Foundation
 /// Helper struct to easily create a `MockUrlSession` that sends the desired HTTP status code and `URLResponse`
 struct MockNetworkStack {
 	var urlSession: MockUrlSession
-	var packageVerifier: SAPDownloadedPackage.Verification
+
 
 	init(
-		mockSession: MockUrlSession,
-		packageVerifier:  @escaping SAPDownloadedPackage.Verification = { _ in return true }
+		mockSession: MockUrlSession
 	) {
 		self.urlSession = mockSession
-		self.packageVerifier = packageVerifier
 	}
 
 	/// Convenience, creates a `MockUrlSession`, `URLResponse` under the hood
@@ -25,7 +23,6 @@ struct MockNetworkStack {
 		httpVersion: String = "HTTP/2",
 		headerFields: [String: String] = [:],
 		responseData: Data?,
-		packageVerifier: @escaping SAPDownloadedPackage.Verification = { _ in return true },
 		requestObserver: MockUrlSession.URLRequestObserver? = nil
 	) {
 		let mockResponse = HTTPURLResponse(
@@ -40,7 +37,7 @@ struct MockNetworkStack {
 			error: nil,
 			urlRequestObserver: requestObserver
 		)
-		self.packageVerifier = packageVerifier
+
 	}
 }
 
@@ -49,7 +46,6 @@ extension HTTPClient {
 	static func makeWith(mock stack: MockNetworkStack) -> HTTPClient {
 		HTTPClient(
 			serverEnvironmentProvider: MockTestStore(),
-			packageVerifier: stack.packageVerifier,
 			session: stack.urlSession
 		)
 	}
