+++ src/xcode/ENA/ENA/Source/Client/__tests__/Mocks/VerificationMock.swift
@@ -1,0 +1,25 @@
+////
+// 🦠 Corona-Warn-App
+//
+
+@testable import ENA
+
+struct MockVerifier: Verify {
+
+	// MARK: - Init
+	
+	// MARK: - Overrides
+	
+	// MARK: - Protocol Verify
+	
+	func verify(_ package: SAPDownloadedPackage) -> Bool {
+		return true
+	}
+	
+	// MARK: - Public
+	
+	// MARK: - Internal
+	
+	// MARK: - Private
+	
+}
