+++ src/xcode/ENA/ENA/Source/Services/Risk/Provider/StatisticsProviding.swift
@@ -16,7 +16,7 @@ protocol StatisticsProviding: class {
 protocol StatisticsFetching {
 	var configuration: HTTPClient.Configuration { get }
 	var session: URLSession { get }
-	var packageVerifier: SAPDownloadedPackage.Verifier { get }
+	var packageVerifier: Verifier { get }
 
 	typealias StatisticsFetchingResultHandler = (Result<StatisticsFetchingResponse, Error>) -> Void
